requirements.txt
annotated-types==0.6.0
anyio==4.3.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
click==8.1.7
colorama==0.4.6
cssselect2==0.7.0
dnspython==2.6.1
email_validator==2.1.1
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.3
fonttools==4.51.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
openpyxl==3.1.2
orjson==3.10.3
pandas==2.2.2
pillow==10.3.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydyf==0.10.0
Pygments==2.18.0
pyphen==0.15.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
reportlab==4.2.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
tinycss2==1.3.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
uvicorn==0.29.0
watchfiles==0.21.0
webencodings==0.5.1
websockets==12.0
zopfli==0.2.3
